Added DTO + Mapper; refactored ParkingLotController;refactored ParkingSlotController

+++ src/main/java/com/dawidcz/parkinglotsystem/model/Payment.java
@@ -18,13 +18,13 @@ public class Payment  implements Serializable {
     @Id
     @GeneratedValue
     private Long id;
-    private int ticketId;
-    private int chargerTickedId;
+    private Long ticketId;
+    private Long chargerTickedId;
     private BigDecimal amount;
     private String paymentMethod;
     private String status;
 
-    public Payment(int ticketId, int chargerTickedId, BigDecimal amount, String paymentMethod, String status) {
+    public Payment(Long ticketId, Long chargerTickedId, BigDecimal amount, String paymentMethod, String status) {
         this.ticketId = ticketId;
         this.chargerTickedId = chargerTickedId;
         this.amount = amount;
@@ -32,7 +32,7 @@ public class Payment  implements Serializable {
         this.status = status;
     }
 
-    public Payment(int ticketId, BigDecimal amount, String paymentMethod, String status) {
+    public Payment(Long ticketId, BigDecimal amount, String paymentMethod, String status) {
         this.ticketId = ticketId;
         this.amount = amount;
         this.paymentMethod = paymentMethod;
